+++ src/components/Hero.tsx
@@ -53,19 +53,19 @@ export default function Hero() {
         <div className="mt-16 grid grid-cols-2 md:grid-cols-4 gap-4 sm:gap-8 text-center">
           <div className="p-2 sm:p-4">
             <p className="text-2xl sm:text-3xl font-bold text-blue-600 dark:text-blue-400">
-              843M+
+              907M+
             </p>
             <p className="text-sm sm:text-base text-gray-600 dark:text-gray-300">Visits</p>
           </div>
           <div className="p-2 sm:p-4">
             <p className="text-2xl sm:text-3xl font-bold text-blue-600 dark:text-blue-400">
-              550K+
+              560K+
             </p>
             <p className="text-sm sm:text-base text-gray-600 dark:text-gray-300">Favorites</p>
           </div>
           <div className="p-2 sm:p-4">
             <p className="text-2xl sm:text-3xl font-bold text-blue-600 dark:text-blue-400">
-              520K
+              493K
             </p>
             <p className="text-sm sm:text-base text-gray-600 dark:text-gray-300">Current Players</p>
           </div>